Fix for change in endpoint
Somehow the Linked Data endpoint changed. The city and street adress are no
longer available in the dataset. They are now OPTIONAL in the search query. If
they happen to return, they are still displayed in the results. If they are not
returned, they are ommited in the results display.